requirements.txt
kafka-python
pyspark
xgboost
scikit-learn
pandas
pyyaml
joblib
pytest
pytest-cov